Fix filtering state on profile, search, explore
